# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "小红书爆款笔记生成 agent Skill",
    "description": "Enterprise-grade hot-note research, multi-model generation, account analytics and controlled publishing Agent Skill.",
    "version": "5.14.5"
  },
  "paths": {
...
    "/v1/accounts/{account_id}/analytics/sync": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Sync Account",
        "operationId": "sync_account_v1_accounts__account_id__analytics_sync_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountAnalytics"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountAnalytics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/accounts/{account_id}/weight/query": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Query Weight",
        "operationId": "query_weight_v1_accounts__account_id__weight_query_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AccountAnalytics"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Analytics"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountWeightReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/accounts/{account_id}/weight": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get Weight",
        "operationId": "get_weight_v1_accounts__account_id__weight_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountWeightReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/accounts/{account_id}/diagnose": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Diagnose Account",
        "operationId": "diagnose_account_v1_accounts__account_id__diagnose_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/xhs_skill__api__routes__accounts__DiagnoseRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Diagnose Account V1 Accounts  Account Id  Diagnose Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/accounts/{account_id}/analytics/browser-sync": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Sync Account From Browser",
        "operationId": "sync_account_from_browser_v1_accounts__account_id__analytics_browser_sync_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountAnalytics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assets": {
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "Upload Asset",
        "operationId": "upload_asset_v1_assets_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_asset_v1_assets_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Upload Asset V1 Assets Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/assets/{asset_id}": {
      "delete": {
        "tags": [
          "assets"
        ],
        "summary": "Delete Asset",
        "operationId": "delete_asset_v1_assets__asset_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Asset V1 Assets  Asset Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/operations/assets": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Search Assets",
        "operationId": "search_assets_v1_operations_assets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Search Assets V1 Operations Assets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountAnalytics": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "followers": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Followers"
          },
          "published_note_count": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Published Note Count"
          },
          "recent_publish_count_7d": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Publish Count 7D"
          },
          "recent_publish_count_30d": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Publish Count 30D"
          },
          "average_publish_interval_hours": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Publish Interval Hours"
          },
          "views_30d": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Views 30D"
          },
          "likes_30d": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Likes 30D"
          },
          "saves_30d": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Saves 30D"
          },
          "comments_30d": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments 30D"
          },
          "shares_30d": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares 30D"
          },
          "follows_gained_30d": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Follows Gained 30D"
          },
          "profile_visits_30d": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Visits 30D"
          },
          "search_views_30d": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Views 30D"
          },
          "recommendation_views_30d": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Views 30D"
          },
          "commercial_note_ratio": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Commercial Note Ratio"
          },
          "deleted_note_count_90d": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted Note Count 90D"
          },
          "violation_count_90d": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Violation Count 90D"
          },
          "latest_violation_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Violation At"
          },
          "category_distribution": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Category Distribution"
          },
          "note_performance": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Note Performance"
          }
        },
        "type": "object",
        "required": [
          "account_id"
        ],
        "title": "AccountAnalytics"
      },
      "AccountWeightReport": {
        "properties": {
          "score_type": {
            "type": "string",
            "title": "Score Type",
            "default": "ESTIMATED_ACCOUNT_WEIGHT"
          },
          "overall_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Score"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "data_completeness": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Data Completeness"
          },
          "dimensions": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DimensionScore"
            },
            "type": "object",
            "title": "Dimensions"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "risks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risks"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Actions"
          },
          "missing_data": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Data"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "OK"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "该分数为系统根据授权数据计算的账号健康与内容分发能力估算，不是小红书官方内部账号权重。"
          }
        },
        "type": "object",
        "required": [
          "confidence",
          "data_completeness"
        ],
        "title": "AccountWeightReport"
      },
...
      "Body_upload_asset_v1_assets_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_asset_v1_assets_post"
      },
...
      "DimensionScore": {
        "properties": {
          "score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "score",
          "weight"
        ],
        "title": "DimensionScore"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "xhs_skill__api__routes__accounts__DiagnoseRequest": {
        "properties": {
          "analytics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountAnalytics"
              },
              {
                "type": "null"
              }
            ]
          },
          "base_topic": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Topic"
          }
        },
        "type": "object",
        "title": "DiagnoseRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}